+++ lib/fs-tree/util.js
@@ -1,0 +1,7 @@
+'use strict';
+
+function byRelativePath(entry) {
+  return entry.relativePath;
+}
+
+module.exports.byRelativePath = byRelativePath;
